backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["SliadnievDanil <sladnevdanil78@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
dj-rest-auth = "^7.0.1"
django-allauth = "^65.3.1"
requests = "^2.32.3"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
